fix: Correct migration condition in Migrator_1_198_0_PR_56
This commit addresses an issue in the migration check condition of Migrator_1_198_0_PR_56. Previously, the condition used the logical OR operator instead of the logical AND operator, causing the migration to not be applied correctly. The correct condition now ensures that the migration is only applied when the database type is either POSTGRES or MYSQL